requirements.txt
asgiref==3.4.1
boto==2.49.0
boto3==1.20.26
botocore==1.23.26
dj-database-url==0.5.0
Django==3.2.10
django-filter==21.1
django-heroku==0.3.1
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
importlib-metadata==4.10.0
jmespath==0.10.0
Markdown==3.3.6
Pillow==8.4.0
psycopg2-binary==2.8.6
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.1
urllib3==1.26.7
whitenoise==5.3.0
zipp==3.6.0
